backend/requirements.txt
psycopg2-binary
sqlalchemy
alembic
pyyaml
requests
pydantic
pgvector
tika
fastapi
uvicorn
